Refactor vxlan_ecmp_utils.py to common
