+++ tests/vxlan/test_vxlan_crm.py
@@ -4,9 +4,7 @@ import ipaddress
 from functools import reduce
 
 from tests.common.helpers.assertions import pytest_assert
-from tests.common.fixtures.ptfhost_utils \
-    import copy_ptftests_directory     # noqa: F401
-from tests.vxlan.vxlan_ecmp_utils import Ecmp_Utils
+from tests.common.vxlan_ecmp_utils import Ecmp_Utils
 from tests.vxlan.test_vxlan_ecmp import (   # noqa: F401
     Test_VxLAN,
     fixture_setUp,
